# micro-services/eventsDisplay/src/main/java/com/cts/events/api/SummaryApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.cts.events.api;

import com.cts.events.model.Summary;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import reactor.core.publisher.Flux;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;


import javax.servlet.http.HttpServletRequest;


import java.util.Optional;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-03-11T17:23:10.822+05:30")

@Api(value = "summary", description = "the summary API")
public interface SummaryApi {

    Logger log = LoggerFactory.getLogger(SummaryApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "", nickname = "summaryAllGet", notes = "get all Summary", response = Summary.class, tags={ "Dashboard", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "fetch", response = Summary.class),
        @ApiResponse(code = 404, message = "No DataFound") })
    @RequestMapping(value = "/summary/all",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default Flux<ResponseEntity<Summary>> summaryAllGet() {
//        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
//            if (getAcceptHeader().get().contains("application/json")) {
//                try {
//                    return new ResponseEntity<>(getObjectMapper().get().readValue("{  \"eventId\" : \"eventId\",  \"pocId\" : 7.061401241503109105224211816675961017608642578125,  \"totalTravelHours\" : 5,  \"project\" : \"project\",  \"livesImpacted\" : 2,  \"pocContact\" : 9.301444243932575517419536481611430644989013671875,  \"overallHours\" : 5,  \"councilName\" : \"councilName\",  \"month\" : \"month\",  \"totalVolunteerHours\" : 1,  \"pocName\" : \"pocName\",  \"baseLocation\" : \"baseLocation\",  \"beneficiaryName\" : \"beneficiaryName\",  \"eventDescription\" : \"eventDescription\",  \"eventName\" : \"eventName\",  \"summaryId\" : 0,  \"totalVolunteers\" : 6,  \"category\" : \"category\",  \"activityType\" : \"activityType\",  \"venueAddress\" : \"venueAddress\",  \"eventDate\" : \"2000-01-23\",  \"status\" : \"status\"}", Summary.class), HttpStatus.NOT_IMPLEMENTED);
//                } catch (IOException e) {
//                    log.error("Couldn't serialize response for content type application/json", e);
//                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
//                }
//            }
//        } else {
//            log.warn("ObjectMapper or HttpServletRequest not configured in default SummaryApi interface so no example is generated");
//        }
        return Flux.just(new ResponseEntity<>(HttpStatus.ACCEPTED));
    }

}